2025.03.13 강휘

+++ elastic/diary_elastic.py
@@ -127,7 +127,7 @@ diary_index_body = {
             "start_date": {"type": "date", "format": "basic_date"},
             "end_date": {"type": "date", "format": "basic_date"},
             "tags": {"type": "keyword"},
-            "total_cost": {"type": "float"},
+            "total_cost": {"type": "Integer"},
             "is_public": {"type": "boolean"},
             "bookmark_count": {"type": "float"},
             "area_code": {"type": "keyword"},
